Run particle list unit test for both core and Cajita
