+++ core/unit_test/particle_list_unit_test.hpp
@@ -1,0 +1,209 @@
+/****************************************************************************
+ * Copyright (c) 2018-2022 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Kokkos_Core.hpp>
+
+#include <Cabana_DeepCopy.hpp>
+#include <Cabana_Fields.hpp>
+#include <Cabana_ParticleList.hpp>
+
+#include <gtest/gtest.h>
+
+namespace Test
+{
+//---------------------------------------------------------------------------//
+// Fields.
+struct CommRank : Cabana::Field::Scalar<int>
+{
+    static std::string label() { return "comm_rank"; }
+};
+
+struct Foo : Cabana::Field::Scalar<double>
+{
+    static std::string label() { return "foo"; }
+};
+
+struct Bar : Cabana::Field::Matrix<double, 3, 3>
+{
+    static std::string label() { return "bar"; }
+};
+
+//---------------------------------------------------------------------------//
+template <class ListType>
+void particleListTest( ListType plist )
+{
+    // Resize the aosoa.
+    auto& aosoa = plist.aosoa();
+    std::size_t num_p = 10;
+    aosoa.resize( num_p );
+    EXPECT_EQ( plist.aosoa().size(), 10 );
+
+    // Populate fields.
+    auto px = plist.slice( Cabana::Field::Position<3>() );
+    auto pm = plist.slice( Foo() );
+    auto pc = plist.slice( CommRank() );
+    auto pf = plist.slice( Bar() );
+
+    Cabana::deep_copy( px, 1.23 );
+    Cabana::deep_copy( pm, 3.3 );
+    Cabana::deep_copy( pc, 5 );
+    Cabana::deep_copy( pf, -1.2 );
+
+    // Check the slices.
+    EXPECT_EQ( px.label(), "position" );
+    EXPECT_EQ( pm.label(), "foo" );
+    EXPECT_EQ( pc.label(), "comm_rank" );
+    EXPECT_EQ( pf.label(), "bar" );
+
+    // Check deep copy.
+    auto aosoa_host =
+        Cabana::create_mirror_view_and_copy( Kokkos::HostSpace(), aosoa );
+    auto px_h = Cabana::slice<0>( aosoa_host );
+    auto pm_h = Cabana::slice<1>( aosoa_host );
+    auto pc_h = Cabana::slice<2>( aosoa_host );
+    auto pf_h = Cabana::slice<3>( aosoa_host );
+    for ( std::size_t p = 0; p < num_p; ++p )
+    {
+        for ( int d = 0; d < 3; ++d )
+            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 );
+
+        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 );
+
+        EXPECT_EQ( pc_h( p ), 5 );
+
+        for ( int i = 0; i < 3; ++i )
+            for ( int j = 0; j < 3; ++j )
+                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 );
+    }
+
+    // Locally modify.
+    Kokkos::parallel_for(
+        "modify", Kokkos::RangePolicy<TEST_EXECSPACE>( 0, num_p ),
+        KOKKOS_LAMBDA( const int p ) {
+            auto particle = plist.getParticle( p );
+
+            for ( int d = 0; d < 3; ++d )
+                get( particle, Cabana::Field::Position<3>(), d ) += p + d;
+
+            get( particle, Foo() ) += p;
+
+            get( particle, CommRank() ) += p;
+
+            for ( int i = 0; i < 3; ++i )
+                for ( int j = 0; j < 3; ++j )
+                    get( particle, Bar(), i, j ) += p + i + j;
+
+            plist.setParticle( particle, p );
+        } );
+
+    // Check the modification.
+    Cabana::deep_copy( aosoa_host, aosoa );
+    for ( std::size_t p = 0; p < num_p; ++p )
+    {
+        for ( int d = 0; d < 3; ++d )
+            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 + p + d );
+
+        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 + p );
+
+        EXPECT_EQ( pc_h( p ), 5 + p );
+
+        for ( int i = 0; i < 3; ++i )
+            for ( int j = 0; j < 3; ++j )
+                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 + p + i + j );
+    }
+}
+
+//---------------------------------------------------------------------------//
+template <class ListType>
+void particleViewTest( ListType plist )
+{
+    // Resize the aosoa.
+    auto& aosoa = plist.aosoa();
+    std::size_t num_p = 10;
+    aosoa.resize( num_p );
+    EXPECT_EQ( plist.aosoa().size(), 10 );
+
+    // Populate fields.
+    auto px = plist.slice( Cabana::Field::Position<3>() );
+    auto pm = plist.slice( Foo() );
+    auto pc = plist.slice( CommRank() );
+    auto pf = plist.slice( Bar() );
+
+    Cabana::deep_copy( px, 1.23 );
+    Cabana::deep_copy( pm, 3.3 );
+    Cabana::deep_copy( pc, 5 );
+    Cabana::deep_copy( pf, -1.2 );
+
+    // Check the slices.
+    EXPECT_EQ( px.label(), "position" );
+    EXPECT_EQ( pm.label(), "foo" );
+    EXPECT_EQ( pc.label(), "comm_rank" );
+    EXPECT_EQ( pf.label(), "bar" );
+
+    // Check deep copy.
+    auto aosoa_host =
+        Cabana::create_mirror_view_and_copy( Kokkos::HostSpace(), aosoa );
+    auto px_h = Cabana::slice<0>( aosoa_host );
+    auto pm_h = Cabana::slice<1>( aosoa_host );
+    auto pc_h = Cabana::slice<2>( aosoa_host );
+    auto pf_h = Cabana::slice<3>( aosoa_host );
+    for ( std::size_t p = 0; p < num_p; ++p )
+    {
+        for ( int d = 0; d < 3; ++d )
+            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 );
+
+        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 );
+
+        EXPECT_EQ( pc_h( p ), 5 );
+
+        for ( int i = 0; i < 3; ++i )
+            for ( int j = 0; j < 3; ++j )
+                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 );
+    }
+
+    // Locally modify.
+    Kokkos::parallel_for(
+        "modify", Kokkos::RangePolicy<TEST_EXECSPACE>( 0, num_p ),
+        KOKKOS_LAMBDA( const int p ) {
+            auto particle = plist.getParticleView( p );
+
+            for ( int d = 0; d < 3; ++d )
+                get( particle, Cabana::Field::Position<3>(), d ) += p + d;
+
+            get( particle, Foo() ) += p;
+
+            get( particle, CommRank() ) += p;
+
+            for ( int i = 0; i < 3; ++i )
+                for ( int j = 0; j < 3; ++j )
+                    get( particle, Bar(), i, j ) += p + i + j;
+        } );
+
+    // Check the modification.
+    Cabana::deep_copy( aosoa_host, aosoa );
+    for ( std::size_t p = 0; p < num_p; ++p )
+    {
+        for ( int d = 0; d < 3; ++d )
+            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 + p + d );
+
+        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 + p );
+
+        EXPECT_EQ( pc_h( p ), 5 + p );
+
+        for ( int i = 0; i < 3; ++i )
+            for ( int j = 0; j < 3; ++j )
+                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 + p + i + j );
+    }
+}
+
+//---------------------------------------------------------------------------//
+
+} // end namespace Test
