+++ core/unit_test/tstParticleList.hpp
@@ -11,216 +11,35 @@
 
 #include <Kokkos_Core.hpp>
 
-#include <Cabana_DeepCopy.hpp>
 #include <Cabana_Fields.hpp>
 #include <Cabana_ParticleList.hpp>
+
+#include <particle_list_unit_test.hpp>
 
 #include <gtest/gtest.h>
 
 namespace Test
 {
 //---------------------------------------------------------------------------//
-// Fields.
-struct CommRank : Cabana::Field::Scalar<int>
-{
-    static std::string label() { return "comm_rank"; }
-};
-
-struct Foo : Cabana::Field::Scalar<double>
-{
-    static std::string label() { return "foo"; }
-};
-
-struct Bar : Cabana::Field::Matrix<double, 3, 3>
-{
-    static std::string label() { return "bar"; }
-};
-
-//---------------------------------------------------------------------------//
-void particleListTest()
-{
-    // Make a particle list.
-    using list_type =
-        Cabana::ParticleList<TEST_MEMSPACE, Cabana::Field::Position<3>, Foo,
-                             CommRank, Bar>;
-    list_type plist( "test_particles" );
-
-    // Resize the aosoa.
-    auto& aosoa = plist.aosoa();
-    std::size_t num_p = 10;
-    aosoa.resize( num_p );
-    EXPECT_EQ( plist.aosoa().size(), 10 );
-
-    // Populate fields.
-    auto px = plist.slice( Cabana::Field::Position<3>() );
-    auto pm = plist.slice( Foo() );
-    auto pc = plist.slice( CommRank() );
-    auto pf = plist.slice( Bar() );
-
-    Cabana::deep_copy( px, 1.23 );
-    Cabana::deep_copy( pm, 3.3 );
-    Cabana::deep_copy( pc, 5 );
-    Cabana::deep_copy( pf, -1.2 );
-
-    // Check the slices.
-    EXPECT_EQ( px.label(), "position" );
-    EXPECT_EQ( pm.label(), "foo" );
-    EXPECT_EQ( pc.label(), "comm_rank" );
-    EXPECT_EQ( pf.label(), "bar" );
-
-    // Check deep copy.
-    auto aosoa_host =
-        Cabana::create_mirror_view_and_copy( Kokkos::HostSpace(), aosoa );
-    auto px_h = Cabana::slice<0>( aosoa_host );
-    auto pm_h = Cabana::slice<1>( aosoa_host );
-    auto pc_h = Cabana::slice<2>( aosoa_host );
-    auto pf_h = Cabana::slice<3>( aosoa_host );
-    for ( std::size_t p = 0; p < num_p; ++p )
-    {
-        for ( int d = 0; d < 3; ++d )
-            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 );
-
-        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 );
-
-        EXPECT_EQ( pc_h( p ), 5 );
-
-        for ( int i = 0; i < 3; ++i )
-            for ( int j = 0; j < 3; ++j )
-                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 );
-    }
-
-    // Locally modify.
-    Kokkos::parallel_for(
-        "modify", Kokkos::RangePolicy<TEST_EXECSPACE>( 0, num_p ),
-        KOKKOS_LAMBDA( const int p ) {
-            auto particle = plist.getParticle( p );
-
-            for ( int d = 0; d < 3; ++d )
-                get( particle, Cabana::Field::Position<3>(), d ) += p + d;
-
-            get( particle, Foo() ) += p;
-
-            get( particle, CommRank() ) += p;
-
-            for ( int i = 0; i < 3; ++i )
-                for ( int j = 0; j < 3; ++j )
-                    get( particle, Bar(), i, j ) += p + i + j;
-
-            plist.setParticle( particle, p );
-        } );
-
-    // Check the modification.
-    Cabana::deep_copy( aosoa_host, aosoa );
-    for ( std::size_t p = 0; p < num_p; ++p )
-    {
-        for ( int d = 0; d < 3; ++d )
-            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 + p + d );
-
-        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 + p );
-
-        EXPECT_EQ( pc_h( p ), 5 + p );
-
-        for ( int i = 0; i < 3; ++i )
-            for ( int j = 0; j < 3; ++j )
-                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 + p + i + j );
-    }
-}
-
-//---------------------------------------------------------------------------//
-void particleViewTest()
-{
-    // Make a particle list.
-    using list_type =
-        Cabana::ParticleList<TEST_MEMSPACE, Cabana::Field::Position<3>, Foo,
-                             CommRank, Bar>;
-
-    list_type plist( "test_particles" );
-
-    // Resize the aosoa.
-    auto& aosoa = plist.aosoa();
-    std::size_t num_p = 10;
-    aosoa.resize( num_p );
-    EXPECT_EQ( plist.aosoa().size(), 10 );
-
-    // Populate fields.
-    auto px = plist.slice( Cabana::Field::Position<3>() );
-    auto pm = plist.slice( Foo() );
-    auto pc = plist.slice( CommRank() );
-    auto pf = plist.slice( Bar() );
-
-    Cabana::deep_copy( px, 1.23 );
-    Cabana::deep_copy( pm, 3.3 );
-    Cabana::deep_copy( pc, 5 );
-    Cabana::deep_copy( pf, -1.2 );
-
-    // Check the slices.
-    EXPECT_EQ( px.label(), "position" );
-    EXPECT_EQ( pm.label(), "foo" );
-    EXPECT_EQ( pc.label(), "comm_rank" );
-    EXPECT_EQ( pf.label(), "bar" );
-
-    // Check deep copy.
-    auto aosoa_host =
-        Cabana::create_mirror_view_and_copy( Kokkos::HostSpace(), aosoa );
-    auto px_h = Cabana::slice<0>( aosoa_host );
-    auto pm_h = Cabana::slice<1>( aosoa_host );
-    auto pc_h = Cabana::slice<2>( aosoa_host );
-    auto pf_h = Cabana::slice<3>( aosoa_host );
-    for ( std::size_t p = 0; p < num_p; ++p )
-    {
-        for ( int d = 0; d < 3; ++d )
-            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 );
-
-        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 );
-
-        EXPECT_EQ( pc_h( p ), 5 );
-
-        for ( int i = 0; i < 3; ++i )
-            for ( int j = 0; j < 3; ++j )
-                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 );
-    }
-
-    // Locally modify.
-    Kokkos::parallel_for(
-        "modify", Kokkos::RangePolicy<TEST_EXECSPACE>( 0, num_p ),
-        KOKKOS_LAMBDA( const int p ) {
-            auto particle = plist.getParticleView( p );
-
-            for ( int d = 0; d < 3; ++d )
-                get( particle, Cabana::Field::Position<3>(), d ) += p + d;
-
-            get( particle, Foo() ) += p;
-
-            get( particle, CommRank() ) += p;
-
-            for ( int i = 0; i < 3; ++i )
-                for ( int j = 0; j < 3; ++j )
-                    get( particle, Bar(), i, j ) += p + i + j;
-        } );
-
-    // Check the modification.
-    Cabana::deep_copy( aosoa_host, aosoa );
-    for ( std::size_t p = 0; p < num_p; ++p )
-    {
-        for ( int d = 0; d < 3; ++d )
-            EXPECT_DOUBLE_EQ( px_h( p, d ), 1.23 + p + d );
-
-        EXPECT_DOUBLE_EQ( pm_h( p ), 3.3 + p );
-
-        EXPECT_EQ( pc_h( p ), 5 + p );
-
-        for ( int i = 0; i < 3; ++i )
-            for ( int j = 0; j < 3; ++j )
-                EXPECT_DOUBLE_EQ( pf_h( p, i, j ), -1.2 + p + i + j );
-    }
-}
-
-//---------------------------------------------------------------------------//
 // RUN TESTS
 //---------------------------------------------------------------------------//
-TEST( TEST_CATEGORY, particle_test ) { particleListTest(); }
+TEST( TEST_CATEGORY, particle_test )
+{
+    Cabana::ParticleList<TEST_MEMSPACE, Cabana::Field::Position<3>, Foo,
+                         CommRank, Bar>
+        plist( "test_particles" );
 
-TEST( TEST_CATEGORY, particle_view_test ) { particleViewTest(); }
+    particleListTest( plist );
+}
+
+TEST( TEST_CATEGORY, particle_view_test )
+{
+    Cabana::ParticleList<TEST_MEMSPACE, Cabana::Field::Position<3>, Foo,
+                         CommRank, Bar>
+        plist( "test_particles" );
+
+    particleViewTest( plist );
+}
 
 //---------------------------------------------------------------------------//
